+++ src/index.ts
@@ -26,7 +26,6 @@ export const ux = {
   get styledHeader() { return deps.styledHeader },
   get styledJSON() { return deps.styledJSON },
   get table() { return deps.table },
-  get supertable() { return deps.supertable },
   get open() { return deps.open },
   get wait() { return deps.wait },
